Added Soundtrack Files, some Dialog Work
