+++ Assets/Materials/PostProcessing/Editor Resources/Monitors.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d1c7c7ad52794a988fdced7348595f52
+folderAsset: yes
+timeCreated: 1523289395
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
