+++ Assets/Materials/PostProcessing/Editor/Attributes.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: c8af2c02fdfd4e189bfd72bdcf2ad9d4
+folderAsset: yes
+timeCreated: 1523289395
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
