+++ Assets/Materials/PostProcessing/Editor/Monitors.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f1a64a65cb19456ebfcfe425fa5a7567
+folderAsset: yes
+timeCreated: 1523289396
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
