+++ Assets/Materials/PostProcessing/Editor/PropertyDrawers.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 5e46cd6585744b2da9135b9ee7c5e300
+folderAsset: yes
+timeCreated: 1523289396
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
