+++ Assets/Materials/PostProcessing/Resources/Bluenoise64.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 71ac76220f1a479c8b1ccb43f76ba06c
+folderAsset: yes
+timeCreated: 1523289395
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
