+++ Assets/Materials/PostProcessing/Resources/Shaders.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d4f223004d6e47e9aa7815f13e62c352
+folderAsset: yes
+timeCreated: 1523289394
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
