+++ Assets/Materials/PostProcessing/Runtime/Attributes.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 63457865160b4499bd38a4e0cec29a4d
+folderAsset: yes
+timeCreated: 1523289393
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
