+++ Assets/Materials/PostProcessing/Runtime/Components.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f202becbf3c54ecca77b9adf74d796b0
+folderAsset: yes
+timeCreated: 1523289393
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
