+++ Assets/Scripts/Dialog/CHDialog.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d4a429d3a0cd4289a2bf875827ae978a
+timeCreated: 1523905947
