+++ Assets/StreamingAssets/Dialog/Ren.json.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 48633f3b2f3dc40bfa6f4908a184daad
+timeCreated: 1523921116
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
