+++ Assets/StreamingAssets/Intro.json
@@ -1,6 +1,8 @@
 {
 	"array": [
-		"You are Magnus Highchurch. You have been sent here to recover the Temporal Chalice.",
+		"You are Magnus Highchurch.",
+		"You used to be a traveling showman, using magic to entertain small audiences.",
+		"You heard the rumors of this town that holds the Temporal Chalice, and were driven to try and recover it."
 		"The Temporal Chalice is an object that can control time itself.",
 		"You are entering a place where time is sick. You cannot exit until you find the chalice and use it to fix the time in this area.",
 		"We are not sure of what you will find inside.",
