+++ Assets/soundtrack/The Clock Strikes Noon (Apocalypse).mp3.meta
@@ -1,0 +1,24 @@
+fileFormatVersion: 2
+guid: d165b34ddcf5d4053814c6f6d818a5d9
+timeCreated: 1523897436
+licenseType: Free
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
