+++ Assets/soundtrack/What's the Plan (Roswell Reprise).mp3.meta
@@ -1,0 +1,24 @@
+fileFormatVersion: 2
+guid: 783380212ba8e4fb68a0c822c34092ae
+timeCreated: 1523897430
+licenseType: Free
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
